+++ shared/contracts/src/index.ts
@@ -3,3 +3,4 @@ export * from './commandStateMachine';
 export * from './sessionStateMachine';
 export * from './transport';
 export * from './actionProgram';
+export * from './aethorArmV1';